requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
twilio==8.5.0
werkzeug==2.3.6